requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
backoff==2.2.1
banks==2.1.3
beautifulsoup4==4.13.4
boto3==1.39.6
botocore==1.39.6
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cohere==5.16.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.5
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
emoji==2.14.1
eval_type_backport==0.2.2
fastapi==0.116.1
fastavro==1.11.1
fastembed==0.7.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.25.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
groq==0.30.0
grpcio==1.73.1
grpcio-status==1.62.3
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langdetect==1.0.9
livekit==1.0.12
livekit-api==1.0.5
livekit-protocol==1.0.4
llama-cloud==0.1.33
llama-cloud-services==0.6.46
llama-index==0.12.49
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.49
llama-index-embeddings-gemini==0.3.2
llama-index-embeddings-google==0.3.1
llama-index-embeddings-google-genai==0.2.1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-instrumentation==0.2.0
llama-index-llms-gemini==0.5.0
llama-index-llms-openai==0.4.7
llama-index-llms-openai-like==0.4.0
llama-index-llms-openrouter==0.3.2
llama-index-multi-modal-llms-openai==0.5.3
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-qdrant==0.6.1
llama-index-workflows==1.1.0
llama-parse==0.6.46
logfire-api==3.24.2
loguru==0.7.3
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.11.0
mdurl==0.1.2
mistralai==1.9.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
olefile==0.47
onnxruntime==1.22.1
openai==1.96.1
opentelemetry-api==1.35.0
packaging==25.0
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.8
portalocker==2.10.1
powerline-status==2.7
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybreaker==1.4.0
pycparser==2.22
pydantic==2.11.7
pydantic-ai==0.4.2
pydantic-ai-slim==0.4.2
pydantic-evals==0.4.2
pydantic-graph==0.4.2
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
pyparsing==3.2.3
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.3
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.4.1
starlette==0.47.1
striprtf==0.0.26
sympy==1.14.0
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.53.2
types-protobuf==6.30.2.20250703
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.18.7
unstructured-client==0.38.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
